Move CurrencySelector to components/*
